refactor(token): move expression classification logic to resolver
- Removed foreign classification code from expression_kind.go
- expression_kind.go now limited to:
  • ExpressionKind enum definitions
  • Basic helpers (String, IsValid, Alias)
- Moved expression parsing and detection logic into resolver.go:
  • Identifier detection
  • Computed expressions
  • Functions and aggregates
  • Subqueries
  • Literals and invalid fallback
- Improves separation of concerns:
  ExpressionKind is declarative, resolver handles classification
